backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain
firecrawl-py
requests
pydantic
python-dotenv
httpx
